fixed double in stock

+++ Client/src/model/Stock.java
@@ -24,7 +24,7 @@ public class Stock implements Serializable {
         this.symbol = symbol;
         this.username = username;
         this.amount = 0;
-        this.price = 0;
+        this.price = 0.0;
     }
 
     /**
@@ -39,15 +39,15 @@ public class Stock implements Serializable {
         this.symbol = symbol;
         this.username = username;
         this.amount = amount;
-        this.price = 0;
+        this.price = 0.0;
     }
 
     /**
      * setting price of stock
      @throws IllegalArgumentException if the user puts negative value
      */
-    public void setPrice(int price) {
-        if (price < 0) throw new IllegalArgumentException();
+    public void setPrice(double price) {
+        if (price < 0.0) throw new IllegalArgumentException();
         this.price = price;
     }
